refactor(feedback): extract reporting module
